Remove HoodieMetadataMergedLogRecordScanner
