+++ hudi-common/src/main/java/org/apache/hudi/metadata/HoodieMetadataLogRecordReader.java
@@ -20,10 +20,7 @@ package org.apache.hudi.metadata;
 
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
-import org.apache.hudi.common.table.log.AbstractHoodieLogRecordReader;
-import org.apache.hudi.common.table.log.BaseHoodieMergedLogRecordScanner;
 import org.apache.hudi.common.table.log.HoodieMergedLogRecordScanner;
-import org.apache.hudi.common.table.log.HoodieMetadataMergedLogRecordScanner;
 import org.apache.hudi.common.table.log.InstantRange;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.collection.ExternalSpillableMap;
@@ -44,8 +41,6 @@ import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import static org.apache.hudi.metadata.HoodieTableMetadataUtil.PARTITION_NAME_SECONDARY_INDEX_PREFIX;
-
 /**
  * Metadata log-block records reading implementation, internally relying on
  * {@link HoodieMergedLogRecordScanner} to merge corresponding Metadata Table's delta log-blocks
@@ -54,17 +49,17 @@ import static org.apache.hudi.metadata.HoodieTableMetadataUtil.PARTITION_NAME_SE
 @ThreadSafe
 public class HoodieMetadataLogRecordReader implements Closeable {
 
-  private final BaseHoodieMergedLogRecordScanner<String> logRecordScanner;
+  private final HoodieMergedLogRecordScanner logRecordScanner;
 
-  private HoodieMetadataLogRecordReader(BaseHoodieMergedLogRecordScanner logRecordScanner) {
+  private HoodieMetadataLogRecordReader(HoodieMergedLogRecordScanner logRecordScanner) {
     this.logRecordScanner = logRecordScanner;
   }
 
   /**
    * Returns the builder for {@code HoodieMetadataMergedLogRecordScanner}.
    */
-  public static HoodieMetadataLogRecordReader.Builder newBuilder(String partitionName) {
-    return new HoodieMetadataLogRecordReader.Builder(partitionName);
+  public static HoodieMetadataLogRecordReader.Builder newBuilder() {
+    return new HoodieMetadataLogRecordReader.Builder();
   }
 
   @SuppressWarnings("unchecked")
@@ -156,20 +151,14 @@ public class HoodieMetadataLogRecordReader implements Closeable {
    * Builder used to build {@code HoodieMetadataMergedLogRecordScanner}.
    */
   public static class Builder {
-    private final AbstractHoodieLogRecordReader.Builder scannerBuilder;
-    private final String partitionName;
-
-    public Builder(String partitionName) {
-      this.partitionName = partitionName;
-      this.scannerBuilder = shouldUseMetadataMergedLogRecordScanner() ? HoodieMetadataMergedLogRecordScanner.newBuilder() : HoodieMergedLogRecordScanner.newBuilder();
-      scannerBuilder
-          .withKeyFieldOverride(HoodieMetadataPayload.KEY_FIELD_NAME)
-          // NOTE: Merging of Metadata Table's records is currently handled using {@code HoodiePreCombineAvroRecordMerger}
-          //       for compatibility purposes; In the future it {@code HoodieMetadataPayload} semantic
-          //       will be migrated to its own custom instance of {@code RecordMerger}
-          .withReverseReader(false)
-          .withOperationField(false);
-    }
+    private final HoodieMergedLogRecordScanner.Builder scannerBuilder =
+        new HoodieMergedLogRecordScanner.Builder()
+            .withKeyFieldOverride(HoodieMetadataPayload.KEY_FIELD_NAME)
+            // NOTE: Merging of Metadata Table's records is currently handled using {@code HoodiePreCombineAvroRecordMerger}
+            //       for compatibility purposes; In the future it {@code HoodieMetadataPayload} semantic
+            //       will be migrated to its own custom instance of {@code RecordMerger}
+            .withReverseReader(false)
+            .withOperationField(false);
 
     public Builder withStorage(HoodieStorage storage) {
       scannerBuilder.withStorage(storage);
@@ -255,11 +244,7 @@ public class HoodieMetadataLogRecordReader implements Closeable {
     }
 
     public HoodieMetadataLogRecordReader build() {
-      return new HoodieMetadataLogRecordReader((BaseHoodieMergedLogRecordScanner) scannerBuilder.build());
-    }
-
-    private boolean shouldUseMetadataMergedLogRecordScanner() {
-      return partitionName.startsWith(PARTITION_NAME_SECONDARY_INDEX_PREFIX);
+      return new HoodieMetadataLogRecordReader(scannerBuilder.build());
     }
   }
 }
